+++ src/main/java/com/tropicoss/alfred/bot/commands/Commands.java
@@ -1,0 +1,16 @@
+package com.tropicoss.alfred.bot.commands;
+
+import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
+import net.dv8tion.jda.api.interactions.commands.build.OptionData;
+
+import java.util.List;
+
+public interface Commands {
+    String getName();
+
+    String getDescription();
+
+    List<OptionData> getOptions();
+
+    void execute(SlashCommandInteractionEvent event);
+}
