+++ src/main/java/com/tropicoss/alfred/config/BotConfig.java
@@ -1,0 +1,6 @@
+package com.tropicoss.alfred.config;
+
+public class BotConfig {
+    public String token = "YourBotToken";
+    public String channel = "YourBotChannel";
+}
