+++ src/main/java/com/tropicoss/alfred/config/GenericConfig.java
@@ -1,0 +1,5 @@
+package com.tropicoss.alfred.config;
+
+public class GenericConfig {
+    public String name = "YourMinecraftServerName";
+}
