+++ src/main/java/com/tropicoss/alfred/config/WebSocketConfig.java
@@ -1,0 +1,13 @@
+package com.tropicoss.alfred.config;
+
+public class WebSocketConfig {
+    public boolean enabled = false;
+    public Type type = Type.SERVER;
+    public String host = "localhost";
+    public int port = 9090;
+
+    public enum Type {
+        SERVER,
+        CLIENT
+    }
+}
