+++ src/main/java/com/tropicoss/alfred/socket/messages/WebsocketMessage.java
@@ -1,0 +1,17 @@
+package com.tropicoss.alfred.socket.messages;
+
+import net.minecraft.text.Text;
+
+public abstract class WebsocketMessage {
+    public String getMessageType() {
+        return null;
+    }
+
+    public String toConsoleString() {
+        return null;
+    }
+
+    public Text toChatText() {
+        return null;
+    }
+}
