--- src/main/java/com/tropicoss/minecraft/alfred/bot/commands/Commands.java
@@ -1,16 +1,0 @@
-package com.tropicoss.minecraft.alfred.bot.commands;
-
-import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
-import net.dv8tion.jda.api.interactions.commands.build.OptionData;
-
-import java.util.List;
-
-public interface Commands {
-    String getName();
-
-    String getDescription();
-
-    List<OptionData> getOptions();
-
-    void execute(SlashCommandInteractionEvent event);
-}
