--- src/main/java/com/tropicoss/minecraft/alfred/config/BotConfig.java
@@ -1,6 +1,0 @@
-package com.tropicoss.minecraft.alfred.config;
-
-public class BotConfig {
-    public String token = "YourBotToken";
-    public String channel = "YourBotChannel";
-}
