--- src/main/java/com/tropicoss/minecraft/alfred/config/GenericConfig.java
@@ -1,5 +1,0 @@
-package com.tropicoss.minecraft.alfred.config;
-
-public class GenericConfig {
-    public String name = "YourMinecraftServerName";
-}
