--- src/main/java/com/tropicoss/minecraft/alfred/socket/messages/WebsocketMessage.java
@@ -1,17 +1,0 @@
-package com.tropicoss.minecraft.alfred.socket.messages;
-
-import net.minecraft.text.Text;
-
-public abstract class WebsocketMessage {
-    public String getMessageType() {
-        return null;
-    }
-
-    public String toConsoleString() {
-        return null;
-    }
-
-    public Text toChatText() {
-        return null;
-    }
-}
