+++ src/main/resources/alfred.mixins.json
@@ -1,7 +1,7 @@
 {
   "required": true,
   "minVersion": "0.8",
-  "package": "com.tropicoss.minecraft.alfred.minecraft.mixin",
+  "package": "com.tropicoss.alfred.minecraft.mixin",
   "compatibilityLevel": "JAVA_17",
   "mixins": [
     "PlayerManagerMixin"
