+++ src/main/resources/fabric.mod.json
@@ -12,7 +12,7 @@
   "environment": "server",
   "entrypoints": {
     "server": [
-      "com.tropicoss.minecraft.alfred.Alfred"
+      "com.tropicoss.alfred.Alfred"
     ]
   },
   "mixins": [
